\name{Quartets}
\alias{Quartets}
\docType{data}
\title{
Anscombe's (1973) Quartets
}
\description{
Four x-y datasets, combined into a long format, which have the same traditional statistical properties (mean, variance, correlation, regression line, etc.) but look quite different.
}
\usage{data("Quartets")}
\format{
  A data frame with 44 observations on the following 3 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
    \item{\code{quartet}}{a character vector for the quartet}
  }
}
\details{

}
\source{

}
\references{
Anscombe, Francis J. (1973). "Graphs in Statistical Analysis." \emph{The American Statistician} 27: 17–21.
}
\examples{
data(Quartets)

}
\keyword{datasets}
