\name{Therms18}
\alias{Therms18}
\docType{data}
\title{
Obama and Trump Thermometer Ratings, 2018
}
\description{
A useful data set (for me) for illustrating central limit theorem, even in ugly data.
}
\usage{data("Therms18")}
\format{
  A data frame with 2500 observations on the following 2 variables.
  \describe{
    \item{\code{ftobama}}{a numeric vector for the thermometer rating for Barack Obama (0-100)}
    \item{\code{fttrump}}{a numeric vector for the thermometer rating for Donald Trump (0-100)}
  }
}
\details{

}
\source{
ANES 2018 pilot study
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(Therms18)
}
\keyword{datasets}
